Updated theme options
