+++ lib/ff_contact_avatar.dart
@@ -3,51 +3,8 @@ library flutter_ff_contact_avatar;
 import 'package:flutter/material.dart';
 import 'dart:math' show min;
 
-class FFContactAvatarTheme {
-  final Color backgroundColor;
-  final Color foregroundColor;
-  final Color badgeColor;
-  final double avatarRadius;
-  final TextStyle initialsTextStyle;
-  final TextStyle nameTextStyle;
-  final TextStyle messageTextStyle;
-  final double verticalPadding;
-  final double avatarElevation;
-
-  FFContactAvatarTheme({
-    this.backgroundColor,
-    this.foregroundColor,
-    this.badgeColor,
-    this.avatarRadius,
-    this.initialsTextStyle,
-    this.nameTextStyle,
-    this.messageTextStyle,
-    this.verticalPadding,
-    this.avatarElevation,
-  });
-
-  static final FFContactAvatarTheme defaultTheme = FFContactAvatarTheme(
-    backgroundColor: Colors.blueGrey,
-    foregroundColor: Colors.white,
-    badgeColor: Colors.red,
-    avatarRadius: 31,
-    initialsTextStyle: TextStyle(
-      fontWeight: FontWeight.bold,
-      fontSize: 26,
-      letterSpacing: 1.4,
-    ),
-    nameTextStyle: TextStyle(
-      fontWeight: FontWeight.bold,
-      fontSize: 13.0,
-    ),
-    messageTextStyle: TextStyle(
-      fontWeight: FontWeight.normal,
-      fontSize: 11.0,
-    ),
-    verticalPadding: 2,
-    avatarElevation: 4,
-  );
-}
+import 'package:ff_contact_avatar/ff-contact-avatar-theme.dart';
+export 'package:ff_contact_avatar/ff-contact-avatar-theme.dart';
 
 // ignore: must_be_immutable
 class FFContactAvatar extends StatelessWidget {
